Update asset-management.js
Added the following to this website page:

Our investment process begins with an analysis of your needs and goals. A profile is developed that highlights your investment objectives, risk tolerance and investment time horizon. Once established, we will work together with you to define an investment plan tailored to your specific situation.